# Conflicts:
#	alicloud/resource_alicloud_db_backup_policy_test.go
#	alicloud/resource_alicloud_gpdb_instance.go
#	alicloud/resource_alicloud_gpdb_instance_test.go
#	alicloud/resource_alicloud_kvstore_instance_test.go
#	alicloud/resource_alicloud_privatelink_vpc_endpoint_service.go
#	alicloud/service_alicloud_polardb.go
#	go.mod

+++ alicloud/resource_alicloud_oss_bucket.go
@@ -1761,6 +1761,9 @@ func resourceAlicloudOssBucketDelete(d *schema.ResourceData, meta interface{}) e
 								VersionId: object.VersionId,
 							})
 						}
+						if len(objectsToDelete) == 0 {
+							return nil, nil
+						}
 						return bucket.DeleteObjectVersions(objectsToDelete)
 					})
 					if er != nil {
